Bootstrapped the h2c.pl script in version 2
The script is working, but there are still some unclear parts, so I will
add comments and reorganize the code in further commits.
It is written with as less dependencies as possible in mind.
So it should work everywhere. I will remove the Data::Dumper dependency
in another commit.

If you want to try it out, read the help page first please.